Separate EE and OSS transforms upsell logic
