+++ frontend/src/metabase/transforms/upsells/components/PythonTransformsUpsellModal/PythonTransformsUpsellModal.tsx
@@ -6,6 +6,7 @@ import { trackUpsellViewed } from "metabase/admin/upsells/components/analytics";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import type { BillingPeriod } from "metabase/data-studio/upsells/types";
 import { useSelector } from "metabase/lib/redux";
+import { PLUGIN_TRANSFORMS } from "metabase/plugins";
 import { getStoreUsers } from "metabase/selectors/store-users";
 import { getIsHosted } from "metabase/setup/selectors";
 import {
@@ -19,9 +20,6 @@ import {
   Title,
 } from "metabase/ui";
 
-import { useTransformsBilling } from "../../hooks";
-
-import { CloudPurchaseContent } from "./CloudPurchaseContent";
 import { SelfHostedContent } from "./SelfHostedContent";
 import { CAMPAIGN, LOCATION } from "./constants";
 
@@ -61,7 +59,7 @@ export function PythonTransformsUpsellModal({
   const modalOpen = isOpen || forceModalToOpen;
 
   const { isLoading, error, billingPeriodMonths, pythonProduct, isOnTrial } =
-    useTransformsBilling();
+    PLUGIN_TRANSFORMS.useTransformsBilling();
 
   useEffect(() => {
     if (isOpen) {
@@ -120,7 +118,7 @@ export function PythonTransformsUpsellModal({
     }
 
     return (
-      <CloudPurchaseContent
+      <PLUGIN_TRANSFORMS.CloudPurchaseContent
         billingPeriod={billingPeriod}
         handleModalClose={handleModalClose}
         isTrialFlow={isTrialFlow}
