+++ frontend/src/metabase/transforms/upsells/components/index.ts
@@ -1,2 +1,1 @@
-export { TransformsSettingUpModal } from "./TransformsSettingUpModal";
 export { PythonTransformsUpsellModal } from "./PythonTransformsUpsellModal";
